+++ main_script.js
@@ -3,15 +3,16 @@
 
 
 // variables
-const version = "v3.4.0-20260318";
+const version = "v3.4.3-20260319";
 const updates = [
         // newest first (preferably, but the program sorts it either way)
         [20291231, "title", "description"], // preset (date title description)
         // shift all (except preset) down when adding new
 
-        [20260321, "AmazonSMP Season 13 Start", "description"],
+        [20260320, "AmazonSMP Season 13 Start", "description"],
         [20260320, "Website Version 4.0 Release", "description"],
-        [20260328, "Website Version 3.4 Beta Release", "description4"],
+        [20260319, "Website Moved To Gitlab", 'The main repository for website project files will now be on Gitlab since <a href="https://gitlab.com/dohst1">Dohst</a> has had issues with github recently. The Github version will still be occasionally updated, although mostly only for major updates or patches.'],
+        [20260318, "Website Version 3.4 Beta Release", "description4"],
         [20260316, "Website Version Renaming", "AmazonSMP Website versions stay as x.y.z (extras, such as x.y.z.v.w, become x.y.(z+v+w)). Dohst Website (rai_website) versions all go from 0.y.z to 3.y.z. Then official release for Dohst Website will be 4.0.0."],
         [20260301, "Dohst23 Type04 Beginning", "From 20260301 to 20280229, supervised by 230023 and 260026, directed by e16, funded by /////////."],
 
@@ -111,10 +112,10 @@ const PageContent = {
     // footer
     content_footer: `<br>
     <p>Website <a href="/main.html#updates">Version ${version}</a>
-    - Written and Maintained by <!-- ///////// and --> <a href="https://github.com/dohst1">Dohst</a></p>
-    <p>Copyright &copy;<a href="https://github.com/dohst1">Dohst</a> 2022-2026, All Rights Reserved.</p>
+    - Written and Maintained by <!-- ///////// and --> <a href="https://gitlab.com/dohst1">Dohst</a></p>
+    <p>Copyright &copy;<a href="https://gitlab.com/dohst1">Dohst</a> 2022-2026, All Rights Reserved.</p>
     <br>`,
-    //<code> &lt;&lt;&lt;<a href="https://www.gov.uk/copyright" target="_blank">&copy;</a><a href="https://github.com/dohst1">2026 Dohst</a> &gt;&gt;&gt; </code>
+    //<code> &lt;&lt;&lt;<a href="https://www.gov.uk/copyright" target="_blank">&copy;</a><a href="https://gitlab.com/dohst1">2026 Dohst</a> &gt;&gt;&gt; </code>
     // ### `<<< © 2026 Dohst >>>`
     set_footer: function set_footer() {document.getElementById("footer").innerHTML = this.content_footer},
     
@@ -303,8 +304,9 @@ const PageContent = {
     // season name, minecraft version, fabric version, fabric api version
     content_amazonsmp_mods_details: ["as13p1", "1.21.11", "0.18.4", "0.141.3"],
     //[0'name', 1'description', 2'version', 3'link'], 
-    content_amazonsmp_mods_required: [[]],
+    //content_amazonsmp_mods_required: [[]],
     content_amazonsmp_mods_all: [['name', 'description', 'version', 'link', 'required'], 
+    ['Fabric API', 'Lightweight and modular API providing common hooks and intercompatibility measures utilized by mods using the Fabric toolchain.', '0.141.3', 'https://modrinth.com/mod/fabric-api/version/0.141.3+1.21.11', 'both', true], 
     ['AmbientSounds', 'listentonature', '6.3.5', 'https://modrinth.com/mod/ambientsounds/version/JZUqW70J', 'client'], 
     ['Architectury API', 'An intermediary api aimed to ease developing multiplatform mods.', '19.0.1', 'https://modrinth.com/mod/architectury-api/version/19.0.1+fabric', 'both'], 
     ['Axiom', 'The all-in-one tool for editing Minecraft Worlds.', '5.3.0', 'https://modrinth.com/mod/axiom/version/uoTNUpOT', 'client'], 
@@ -317,33 +319,95 @@ const PageContent = {
     ['CreativeCore', 'A core mod', '2.14.11', 'https://modrinth.com/mod/creativecore/version/2.14.11', 'client'], 
     ['Distant Horizons', 'See farther without turning your game into a slide show', '2.4.5-b', 'https://modrinth.com/mod/distanthorizons/version/2.4.5-b-1.21.11', 'client'], 
     ['Entity Culling', 'Using async path-tracing to hide Block-/Entities that are not visible', '1.9.5', 'https://modrinth.com/mod/entityculling/version/Dx3xsUER', 'client'], 
-    ['Fabric API', 'Lightweight and modular API providing common hooks and intercompatibility measures utilized by mods using the Fabric toolchain.', '0.141.3', 'https://modrinth.com/mod/fabric-api/version/0.141.3+1.21.11', 'both'], 
     ['FerriteCore', 'Memory usage optimizations.', '8.2.0', 'https://modrinth.com/mod/ferrite-core/version/8.2.0-fabric', 'both'], 
     ['Floodgate fabric', 'Hybrid mode plugin to allow for connections from Geyser to join online mode servers.', '2.2.6-b54', 'https://modrinth.com/mod/floodgate/version/wzwExuYr', 'server'],
     ['Freecam', 'A highly customizable freecam mod.', '1.3.6', 'https://modrinth.com/mod/freecam/version/1.3.6+mc1.21.11', 'client'], 
     ['GeyserMC fabric', 'Enable clients from Minecraft Bedrock Edition to join your Minecraft Java server.', 'Build 1099 14/03/2026', 'https://geysermc.org', 'server'], 
     ['Iris Shaders', 'A modern shader pack loader for Minecraft intended to be compatible with existing OptiFine shader packs', '1.10.6', 'https://modrinth.com/mod/iris/version/1.10.6+1.21.11-fabric', 'client'], 
     ['Litematica', 'A client-side schematic mod with extra features for creative mode work', '0.26.1', 'https://modrinth.com/mod/litematica/version/0.26.1', 'client'], 
-    ['Lithium', 'No-compromises game logic optimization mod. Well suited for clients and servers of all kinds. Now available for Fabric and NeoForge!', '0.21.4', 'https://modrinth.com/mod/lithium/version/mc1.21.11-0.21.4-fabric', 'both'], 
+    ['Lithium', 'No-compromises game logic optimization mod. Well suited for clients and servers of all kinds. Now available for Fabric and NeoForge!', '0.21.4', 'https://modrinth.com/mod/lithium/version/mc1.21.11-0.21.4-fabric', 'both', true], 
     ['MaLiLib', "A library mod for client-side mods. Contains most of the common/shared code of masa's client mods, and adds some inter-operation support between the dependent mods.", '0.27.7', 'https://modrinth.com/mod/malilib/version/0.27.7', 'client'], 
     ['MiniHUD', "A 'mini F3' HUD mod, also with various overlays like light level, spawn chunks, slime chunks etc.", '0.38.4', 'https://modrinth.com/mod/minihud/version/0.38.4', 'client'], 
     ['Mod Menu', 'Adds a mod menu to view the list of mods you have installed.', 'v17.0.0-beta.2', 'https://modrinth.com/mod/modmenu/version/17.0.0-beta.2', 'client'], 
-    ['No Chat Reports', 'Makes chat unreportable (where possible)', '2.18.0', 'https://modrinth.com/mod/no-chat-reports/version/Fabric-1.21.11-v2.18.0', 'both'], 
-    ['No Chat Restrictions', 'Restores access to game chat for all accounts', '1.0.0', 'https://modrinth.com/mod/no-chat-restrictions/version/Fabric-MC1.21.11-v1.0.0', 'client'], 
+    ['No Chat Reports', 'Makes chat unreportable (where possible)', '2.18.0', 'https://modrinth.com/mod/no-chat-reports/version/Fabric-1.21.11-v2.18.0', 'both', true], 
+    ['No Chat Restrictions', 'Restores access to game chat for all accounts', '1.0.0', 'https://modrinth.com/mod/no-chat-restrictions/version/Fabric-MC1.21.11-v1.0.0', 'client', true], 
     ['ReplayMod', 'A Minecraft Mod to record, relive and share your experience.', '2.6.25', 'https://modrinth.com/mod/replaymod/version/1.21.11-2.6.25', 'client'], 
     ['Replay Voice Chat', 'A compatibility mod to record Simple Voice Chat with ReplayMod', '1.3.12', 'https://modrinth.com/mod/replay-voice-chat/version/fabric-1.21.11-1.3.12', 'client'], 
-    ['Roughly Enough Items (REI)', 'Clean and Customizable. Alternative to Just Enough Items/JEI.', '21.11.814', 'https://modrinth.com/mod/rei/version/21.11.814+fabric', 'both'], 
+    ['Roughly Enough Items (REI)', 'Clean and Customizable. Alternative to Just Enough Items/JEI.', '21.11.814', 'https://modrinth.com/mod/rei/version/21.11.814+fabric', 'both', true], 
     ['ScalableLux', 'A Fabric mod based on Starlight that improves the performance of light updates in Minecraft.', '0.1.6', 'https://modrinth.com/mod/scalablelux/version/0.1.6+fabric.c25518a', 'both'], 
     ['Servux', 'Servux is a server-side mod that provides support for some client-side mods, such as sending structure bounding boxes for MiniHUD', '0.9.2', 'https://modrinth.com/mod/servux/version/0.9.2', 'server'], 
     ['Simple Voice Chat', 'A working voice chat in Minecraft.', '2.6.12', 'https://modrinth.com/plugin/simple-voice-chat/version/fabric-1.21.11-2.6.12', 'both'], 
-    ['Sodium', 'The fastest and most compatible rendering optimization mod for Minecraft. Now available for both NeoForge and Fabric!', '0.8.6', 'https://modrinth.com/mod/sodium/version/mc1.21.11-0.8.6-fabric', 'client'], 
+    ['Sodium', 'The fastest and most compatible rendering optimization mod for Minecraft. Now available for both NeoForge and Fabric!', '0.8.6', 'https://modrinth.com/mod/sodium/version/mc1.21.11-0.8.6-fabric', 'client', true], 
     ['TCDCommons API', "TheCSDev's personal library mod for the Minecraft modding environment.", '5.0.0-beta.6', 'https://modrinth.com/mod/tcdcommons/version/5.0.0-beta.6+fabric-1.21.11', 'client'], 
     ['Text Placeholder API', 'Placeholder and Text manipulation library for your Minecraft mods.', '2.8.2', 'https://modrinth.com/mod/placeholder-api/version/2.8.2+1.21.10', 'client'], 
-    ["Xaero's Minimap", "Displays a map of the nearby world terrain, players, mobs, entities in the corner of your screen. Lets you create waypoints which help you find the locations you've marked.", '25.3.10', 'https://modrinth.com/mod/xaeros-minimap/version/fabric-1.21.11-25.3.10', 'client'], 
-    ["Xaero's World Map", "Adds a full screen world map which shows you what you have explored in the world. Works great together with Xaero's Minimap.", '1.40.11', 'https://modrinth.com/mod/xaeros-world-map/version/fabric-1.21.11-1.40.11', 'client'], 
+    ["Xaero's Minimap", "Displays a map of the nearby world terrain, players, mobs, entities in the corner of your screen. Lets you create waypoints which help you find the locations you've marked.", '25.3.10', 'https://modrinth.com/mod/xaeros-minimap/version/fabric-1.21.11-25.3.10', 'client', true], 
+    ["Xaero's World Map", "Adds a full screen world map which shows you what you have explored in the world. Works great together with Xaero's Minimap.", '1.40.11', 'https://modrinth.com/mod/xaeros-world-map/version/fabric-1.21.11-1.40.11', 'client', true], 
     ],
-    content_amazonsmp_datapacks: [[]],
-    set_amazonsmp: function (target, option = "all") {
+    // content_amazonsmp_datapacks, 35, 4, [name, description, author, link]
+    content_amazonsmp_datapacks: [['name', 'description', 'author', 'link'], 
+                               ['afk display', "Changes a player's name color if they havent moved for over 5 minutes.", '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['age lock', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['armor statues', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['armored elytra', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['cauldron concrete', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['cauldron mud', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['classic fishing loot', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['coordinates hud', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['custom nether portals', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['double shulker shells', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['dragon drops', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['durability ping', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['glass always drops', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['husks drop sand', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['mini blocks', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['more effective tools', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['more mob heads', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['name colors', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['nether portal coords', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['no ender eye (v2.1)', 'Makes eyes of ender uncraftable.', '<a href="https://gitlab.com/dohst1">Dohst</a>', 'dohst.html#datapacks'], 
+                               ['painting picker', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['player head drops', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['silence mobs', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['silk touch budding amethyst', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['spectator conduit power', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['spectator night vision', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['storm channeling', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['track raw statistics', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['unlock all recipes', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['crafting tweaks', 'Everything except rabbit hide buddles and three by three coral blocks.', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/crafting-tweaks/'], 
+                               ['villager death messages', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['wandering trader announcements', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['wandering trades', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['xp bottling', '', '<a href="https://vanillatweaks.net">VanillaTweaks</a>', 'https://vanillatweaks.net/picker/datapacks/'], 
+                               ['amplified nether', 'The nether explorers simple dream, doubled height and amplified terrain.', '<a href="https://www.stardustlabs.net">Stardust Labs</a>', 'https://modrinth.com/datapack/amplified-nether/version/oZdWnGL8']],
+
+    content_amazonsmp_players: [['name', 'username', 'uuid', 'admin', 'banned', 'link'], 
+                             ['Dohst', 'Dohst0', 'a72d1775-1031-4937-9fcb-3296809ef742', 'yes', 'no', 'https://gitlab.com/dohst1'], 
+                             ['Secret7', 'Secret7', '73045442-ec92-419c-a10c-d8efdab5f555', 'yes', 'no', ''], 
+                             ['Cookielotty12', 'Cookielotty12', '193df712-49dd-4bd1-a21d-09cc6f953bd9', 'no', 'no', ''], 
+                             ['Tackzs', 'Tackzs', 'd105fc83-6c49-4bfe-b8a3-0735cee598b4', 'no', 'no', ''], 
+                             ['WhatCheeseburger', 'WhatCheeseburger', '438f0db2-ad4f-44d8-8d21-2bbb5600e676', 'yes', 'no', 'https://www.twitch.tv/whatcheeseburgertv'], 
+                             ['MandoEAM', 'MandoEAM', '9df74492-2311-483f-9f8b-4ba35d9af733', 'yes', 'no', 'https://www.youtube.com/@MandoEAM'], 
+                             ['ATAT66', 'ATAT66', '0ca533c1-e6a6-486a-86d0-aabe670031b4', 'no', 'no', ''], 
+                             ['GoGapplejuice', 'GoGapplejuice', 'c50ab55c-a3b1-42e5-b2d5-772ddc3f5584', 'no', 'no', ''], 
+                             ['Dannygpr', '.Dannygpr00981', '00000000-0000-0000-0009-01f66bfb89a7', 'no', 'no', 'https://youtube.com/@danny_gpr'], 
+                             ['Filterrs', 'Filterrs', 'b0c9a712-0db0-4f23-9e25-128e259f8223', 'no', 'no', ''], 
+                             /* ['230023', 'unknown', 'unknown', '0', 'yes'] */],
+                             /* 
+                            <tr>
+                            <th>Dohst_</th>
+                            <th>414ccf90-f3cb-4c38-a3b1-ebeacb4f5882</th>
+                            <th><a href="https://gitlab.com/dohst1">Dohst</a></th>
+                            <th>yes</th> <!-- A3 -->
+                        </tr>   
+                             <tr>
+                            <th>EndaDwagon</th>
+                            <th>3465711c-9766-4f21-87ff-491155d7a591</th>
+                            <th><a href="https://www.youtube.com/@endadwagon">Enda</a></th>
+                            <th>no</th>
+                        </tr> */
+
+    set_amazonsmp: function (target_mods, target_players, option = "all") {
         // old
         /* if (index == 0 || index >= updates.length || !index) {return null}
         else if (description) {document.getElementById(target).innerHTML = updates[index][2]} 
@@ -351,13 +415,11 @@ const PageContent = {
         
         if (option == "all") {
             let text = `
-                <h2>Mods</h2>
-
-                <h3>Season 13</h3>
+                <h2>Mods and Datapacks</h2>
                 <table>
                 <thead>
                     <tr>
-                        <th colspan="3">Required Mods (minecraft ${this.content_amazonsmp_mods_details[1]}, fabric ${this.content_amazonsmp_mods_details[2]})</th>
+                        <th colspan="3">Recommended Mods (minecraft ${this.content_amazonsmp_mods_details[1]}, fabric ${this.content_amazonsmp_mods_details[2]})</th>
                     </tr>
                     <tr>
                         <th>Name</th>
@@ -366,24 +428,22 @@ const PageContent = {
                     </tr>
                 </thead>
                 <tbody>`;
-            /* for (let i = 1; i < this.content_amazonsmp_mods_required.length; i++) {
+            for (let i = 1; i < this.content_amazonsmp_mods_all.length; i++) {
+                if (this.content_amazonsmp_mods_all[i][5]) {
                 //[0'name', 1'description', 2'version', 3'link'], 
-                text += `<tr>"
-                            <td><a href="${this.content_amazonsmp_mods_required[i][3]}">${this.content_amazonsmp_mods_required[i][0]}</a></td>
-                            <td>${this.content_amazonsmp_mods_required[i][2]}</td>
-                            <td class="limited_big">${this.content_amazonsmp_mods_required[i][1]}</td>
+                text += `<tr>
+                            <td><a href="${this.content_amazonsmp_mods_all[i][3]}">${this.content_amazonsmp_mods_all[i][0]}</a></td>
+                            <td>${this.content_amazonsmp_mods_all[i][2]}</td>
+                            <td class="limited_big">${this.content_amazonsmp_mods_all[i][1]}</td>
                         </tr>`;
-            }; */
-            text += `<tr>
-                <td colspan="3">There are no required mods this season, you can join with a vanilla client.</>
-            </tr>`
+            }};
             text += `</tbody>
             </table>`;
 
             text += `<table>
                     <thead>
                         <tr>
-                            <th colspan="3">All Mods (minecraft ${this.content_amazonsmp_mods_details[1]}, fabric ${this.content_amazonsmp_mods_details[2]})</th>
+                            <th colspan="3">Server Mods (minecraft ${this.content_amazonsmp_mods_details[1]}, fabric ${this.content_amazonsmp_mods_details[2]})</th>
                         </tr>
                         <tr>
                             <th>Name</th>
@@ -393,20 +453,21 @@ const PageContent = {
                     </thead>
                 <tbody>`
             for (let i = 1; i < this.content_amazonsmp_mods_all.length; i++) {
+                if (this.content_amazonsmp_mods_all[i][4] != "client") {
                 //[0'name', 1'description', 2'version', 3'link'], 
                 text += `<tr>
                             <td><a href="${this.content_amazonsmp_mods_all[i][3]}">${this.content_amazonsmp_mods_all[i][0]}</a></td>
                             <td>${this.content_amazonsmp_mods_all[i][2]}</td>
                             <td class="limited_big">${this.content_amazonsmp_mods_all[i][1]}</td>
                         </tr>`;
-            };
+            }};
             text += `</tbody>
             </table>`;
 
             text += `<table>
                     <thead>
                         <tr>
-                            <th colspan="3">Datapacks (minecraft ${this.content_amazonsmp_mods_details[1]})</th>
+                            <th colspan="3">Server Datapacks (minecraft ${this.content_amazonsmp_mods_details[1]})</th>
                         </tr>
                         <tr>
                             <th>Name</th>
@@ -417,7 +478,7 @@ const PageContent = {
                 <tbody>`
             for (let i = 1; i < this.content_amazonsmp_datapacks.length; i++) {
                 //[0'name', 1'description', 2'version', 3'link'], 
-                text += `<tr>"
+                text += `<tr>
                             <td><a href="${this.content_amazonsmp_datapacks[i][3]}">${this.content_amazonsmp_datapacks[i][0]}</a></td>
                             <td>${this.content_amazonsmp_datapacks[i][2]}</td>
                             <td class="limited_big">${this.content_amazonsmp_datapacks[i][1]}</td>
@@ -425,7 +486,44 @@ const PageContent = {
             };
             text += `</tbody></table>`;
 
-            document.getElementById(target).innerHTML = text;
+
+            let text_players = `
+            <h2>Players</h2>
+                <table>
+                    <thead>
+                        <tr>
+                            <th>Name</th>
+                            <th>Username</th>
+                            <!-- <th>UUID</th> -->
+                            <th>Admin</th>
+                            <th>Banned</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+            `
+            for (let i = 1; i < this.content_amazonsmp_players.length; i++) {
+                //[0'name', 1'description', 2'version', 3'link'], 
+                if (this.content_amazonsmp_players[i][0] == this.content_amazonsmp_players[i][1]) {
+                    text_players += `<tr>
+                            <td colspan="2"><a href="${this.content_amazonsmp_players[i][5]}">${this.content_amazonsmp_players[i][0]}</a></td>
+                            <!-- <td>${this.content_amazonsmp_players[i][2]}</td> -->
+                            <td>${this.content_amazonsmp_players[i][3]}</td>
+                            <td>${this.content_amazonsmp_players[i][4]}</td>
+                        </tr>`;
+                } else {
+                    text_players += `<tr>
+                            <td><a href="${this.content_amazonsmp_players[i][5]}">${this.content_amazonsmp_players[i][0]}</a></td> 
+                            <td>${this.content_amazonsmp_players[i][1]}</td>
+                            <!-- <td>${this.content_amazonsmp_players[i][2]}</td> -->
+                            <td>${this.content_amazonsmp_players[i][3]}</td>
+                            <td>${this.content_amazonsmp_players[i][4]}</td>
+                        </tr>`;
+                }
+            };
+            text_players += `</tbody></table>`;
+
+            document.getElementById(target_mods).innerHTML = text;
+            document.getElementById(target_players).innerHTML = text_players;
         };